refactor(face): centralize keyboard scroll handling
- Toggle keyboard-adjusted class inside Face when mobile keyboard is visible
- Keep chat/contact components free of keyboard-specific wrappers
- Provide base face-content utility to enable scrolling without shrinking cube
